Broke TestUtil and subclasses out into separate files; renamed NistHypoExamples to NistTA3Examples.
