+++ src/test/java/com/ncc/aif/TestUtils.java
@@ -1,6 +1,5 @@
 package com.ncc.aif;
 
-import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.apache.jena.rdf.model.Resource;
@@ -297,188 +296,5 @@ class TestUtils {
             }
             fail("Validation was expected to " + (expected ? "pass" : "fail") + " but did not");
         }
-    }
-}
-
-/**
- * An extension of the TestUtils that supports testing TA1 and TA2 restricted AIF.
- * Usage and features are the same as TestUtils.
- * <p>
- * Note that in many cases, you can call makeValidAIFXXX() to make a valid <i>unrestricted</i> AIF object
- * for use in invalid tests, whereas makeValidNistXXX() always returns a valid restricted AIF object.
- */
-class NistTestUtils extends TestUtils {
-
-    /**
-     * Constructor for utilities for testing restricted AIF functionality.
-     *
-     * @param annotationNamespace namespace to use with URIs
-     * @param validator           an AIF validator instantiated based on the caller's ontology and desired NIST restrictions
-     * @param forceDump           whether or not to force dumping of models prior to validation
-     */
-    NistTestUtils(String annotationNamespace, ValidateAIF validator, boolean forceDump) {
-        super(annotationNamespace, validator, forceDump);
-    }
-
-    /**
-     * Call before each test.  Returns a new, empty model with standard AIF namespaces.
-     *
-     * @return a new model with which to start a test; caller may wish to add prefixes for the ontology and annotation
-     */
-    Model startNewTest() {
-        Model model = super.startNewTest();
-        // NIST tests always need type assertion justifications
-        return model;
-    }
-
-    @Override
-    Resource makeValidJustification() {
-        return makeValidJustification(getDocumentName());
-    }
-
-    Resource makeValidJustification(String sourceDocument) {
-        Resource justification = super.makeValidJustification();
-        addSourceDocumentToJustification(justification, sourceDocument);
-        return justification;
-    }
-
-    /**
-     * Makes and returns a valid NIST-restricted entity of the specified type and its cluster.
-     *
-     * @param type entity type
-     * @return a key-value Pair of the entity Resource (key) and its associated cluster Resource (value)
-     */
-    ImmutablePair<Resource, Resource> makeValidNistEntity(Resource type) {
-        return makeValidNistObject(type, makeEntity(model, getEntityUri(), system));
-    }
-
-    /**
-     * Makes and returns a valid NIST-restricted event of the specified type and its cluster.
-     *
-     * @param type event type
-     * @return a key-value Pair of the event Resource (key) and its associated cluster Resource (value)
-     */
-    ImmutablePair<Resource, Resource> makeValidNistEvent(Resource type) {
-        return makeValidNistObject(type, makeEvent(model, getEventUri(), system));
-    }
-
-    /**
-     * Makes and returns a valid NIST-restricted relation of the specified type and its cluster.
-     *
-     * @param type relation type
-     * @return a key-value Pair of the relation Resource (key) and its associated cluster Resource (value)
-     */
-    ImmutablePair<Resource, Resource> makeValidNistRelation(Resource type) {
-        return makeValidNistObject(type, makeRelation(model, getEventUri(), system));
-    }
-
-    // Helper function for makeValidNistXXX
-    private ImmutablePair<Resource, Resource> makeValidNistObject(Resource type, Resource object) {
-        markJustification(addType(object, type), makeValidJustification());
-        Resource cluster = makeClusterWithPrototype(model, getClusterUri(), object, system);
-        return new ImmutablePair<>(object, cluster);
-    }
-}
-
-
-/**
- * An extension of the NistTestUtils that supports testing TA3 restricted AIF.
- * Usage and features are the same as NistTestUtils.
- * <p>
- * Note that in many cases, you can call makeValidAIFXXX() to make a valid <i>unrestricted</i> AIF object
- * or makeValidNistXXX() to make a valid TA1/TA2 restricted AIF object for use in invalid tests,
- * whereas makeValidNistTA3XXX() always returns a valid restricted AIF object for use in TA3.
- */
-class NistTA3TestUtils extends NistTestUtils {
-
-    /**
-     * Constructor for utilities for testing TA3 restricted AIF functionality.
-     *
-     * @param annotationNamespace namespace to use with URIs
-     * @param validator           an AIF validator instantiated based on the caller's ontology and desired NIST restrictions
-     * @param forceDump           whether or not to force dumping of models prior to validation
-     */
-    NistTA3TestUtils(String annotationNamespace, ValidateAIF validator, boolean forceDump) {
-        super(annotationNamespace, validator, forceDump);
-    }
-
-    /**
-     * Makes and returns a valid TA3 NIST-restricted entity of the specified type and its cluster with the specified
-     * cluster handle.
-     *
-     * @param type          entity type
-     * @param clusterHandle cluster handle for the entity cluster
-     * @return a key-value Pair of the entity Resource (key) and its associated cluster Resource (value)
-     */
-    ImmutablePair<Resource, Resource> makeValidNistTA3Entity(Resource type, String clusterHandle) {
-        ImmutablePair<Resource, Resource> pair = makeValidNistEntity(type);
-        pair.getValue().addProperty(AidaAnnotationOntology.HANDLE, clusterHandle);
-        return pair;
-    }
-
-    /**
-     * Makes and returns a valid TA3 NIST-restricted event of the specified type and its cluster marked with the
-     * specified importance.
-     *
-     * @param type       event type
-     * @param importance the importance to mark the event cluster
-     * @return a key-value Pair of the event Resource (key) and its associated cluster Resource (value)
-     */
-    ImmutablePair<Resource, Resource> makeValidNistTA3Event(Resource type, double importance) {
-        ImmutablePair<Resource, Resource> pair = makeValidNistEvent(type);
-        markImportance(pair.getValue(), importance);
-        return pair;
-    }
-
-    /**
-     * Makes and returns a valid TA3 NIST-restricted relation of the specified type and its cluster marked with the
-     * specified importance.
-     *
-     * @param type       relation type
-     * @param importance the importance to mark the relation cluster
-     * @return a key-value Pair of the event Resource (key) and its associated cluster Resource (value)
-     */
-    ImmutablePair<Resource, Resource> makeValidNistTA3Relation(Resource type, double importance) {
-        ImmutablePair<Resource, Resource> pair = makeValidNistRelation(type);
-        markImportance(pair.getValue(), importance);
-        return pair;
-    }
-
-    /**
-     * Makes and returns a valid TA3 argument assertion between the specified event or relation and an argument filler entity.
-     *
-     * @param eventOrRelation The event or relation for which to mark the specified argument role
-     * @param type            the type of the argument
-     * @param argumentFiller  the filler (object) of the argument
-     * @param importance      the importance to mark the edge
-     * @return the created event or relation argument assertion
-     */
-    Resource makeValidTA3Edge(Resource eventOrRelation, Resource type, Resource argumentFiller, double importance) {
-        Resource edge = makeValidAIFEdge(eventOrRelation, type, argumentFiller);
-        markImportance(edge, importance);
-        return edge;
-    }
-
-    /**
-     * Makes and returns a valid TA3 NIST-restricted hypothesis involving the specified resource(s).
-     *
-     * @param resources A set of entities, relations, and arguments that contribute to the hypothesis
-     */
-    Resource makeValidTA3Hypothesis(Resource... resources) {
-        Resource hypothesis = makeValidAIFHypothesis(resources);
-        markImportance(hypothesis, 100.0);
-        return hypothesis;
-    }
-
-    /**
-     * Makes and returns a valid TA3 NIST-restricted hypothesis involving the specified resource(s) and importance.
-     *
-     * @param importance the importance with which to make the hypothesis
-     * @param resources  A set of entities, relations, and arguments that contribute to the hypothesis
-     */
-    Resource makeValidTA3Hypothesis(double importance, Resource... resources) {
-        Resource hypothesis = makeValidAIFHypothesis(resources);
-        markImportance(hypothesis, importance);
-        return hypothesis;
     }
 }
